@@ -12,24 +12,24 @@
     "test": "ng test --code-coverage"
   },
   "dependencies": {
-    "@angular/common": "2.3.1",
-    "@angular/compiler": "2.3.1",
-    "@angular/compiler-cli": "2.3.1",
-    "@angular/core": "2.3.1",
-    "@angular/forms": "2.3.1",
-    "@angular/http": "2.3.1",
-    "@angular/platform-browser": "2.3.1",
-    "@angular/platform-browser-dynamic": "2.3.1",
-    "@angular/platform-server": "2.3.1",
+    "@angular/common": "2.4.8",
+    "@angular/compiler": "2.4.8",
+    "@angular/compiler-cli": "2.4.8",
+    "@angular/core": "2.4.8",
+    "@angular/forms": "2.4.8",
+    "@angular/http": "2.4.8",
+    "@angular/platform-browser": "2.4.8",
+    "@angular/platform-browser-dynamic": "2.4.8",
+    "@angular/platform-server": "2.4.8",
     "@ionic/storage": "1.1.7",
     "@ngrx/core": "^1.2.0",
     "@ngrx/store": "^2.2.1",
     "@ngrx/effects": "^2.0.0",
-    "ionic-angular": "2.0.1",
+    "ionic-angular": "2.3.0",
     "ionic-native": "2.4.1",
     "ionicons": "3.0.0",
     "moment": "^2.16.0",
-    "rxjs": "5.0.0-rc.4",
+    "rxjs": "5.0.1",
     "zone.js": "0.7.2",
     "sw-toolbox": "3.4.0"
   },
